Create user from facebook account.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->increments('id');
             $table->integer('vk_id')->nullable();
             $table->integer('ok_id')->nullable();
-            $table->integer('facebook_id')->nullable();
+            $table->bigInteger('facebook_id')->nullable();
             $table->string('username')->unique();
             $table->string('banner')->nullable();
             $table->string('avatar')->nullable();
